Setup basic input form and seperate requests/responses
