--- frontend/src/data/main_command.ts
@@ -1,8 +1,0 @@
-export interface MainCommandResponse {
-  id: number;
-  name: string
-  params: string | null;
-  format: string | null;
-  data_size: number;
-  total_size: number;
-}
