+++ frontend/src/data/request.ts
@@ -1,0 +1,6 @@
+export interface CommandRequest {
+  name: string
+  params: string | null
+  format: string | null
+}
+
