+++ frontend/src/input/command_input.css
@@ -1,0 +1,7 @@
+.spreader {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+}
