+++ frontend/src/input/command_input.tsx
@@ -1,14 +1,27 @@
+import "./command_input.css"
+
 const CommandInput = () => {
+  // Setup state and useEffect calls here
+
+  const handleSubmit = () => {
+    // TODO:(Member) Submit to your post endpoint 
+  }
+
   return (
     <>
       <form>
-        <label>Command Type: </label>
-        <select>
-          <option value="1">Command 1</option>
-          <option value="2">Command 2</option>
-          <option value="3">Command 3</option>
-        </select>
-        <button type="submit">Submit</button>
+        <div className="spreader">
+          <div>
+            <label>Command Type: </label>
+            <select>{/*Display the list of commands based on the get commands request*/}
+              <option value={"1"}>Command 1</option>
+              <option value={"2"}>Command 2</option>
+              <option value={"3"}>Command 3</option>
+            </select>
+          </div>
+          <input /> {/*Add input handling here if the selected command has a param input*/}
+          <button onClick={handleSubmit}>Submit</button>
+        </div>
       </form>
     </>
   )
